More advanced deployment check
